Refactor spy program to allow code sharing
